fix(playerbot): Move GuildMetrics to interface, remove GuildEventCoordinator dup destructor
- Moved GuildMetrics struct to IGuildIntegration.h with atomic members, copy constructor, and assignment operator
- Removed duplicate structs from GuildIntegration.h
- Fixed GuildEventCoordinator duplicate destructor

Next: Fix EventCoordinationProfile override error (same pattern - move struct to interface)
